Upload default beep and use instead of error (#160)
* Upload default beep and use instead of error

* Break get beep once found

* Update ProfanityRemover to Censor agent with way for user to give prompt

* Improve beep precision and handle small words edge case
